+++ src/main/java/drjoliv/jfunc/io/IO.java
@@ -1,14 +1,13 @@
 package drjoliv.jfunc.io;
 
 import java.io.FileReader;
+import java.io.InputStream;
 import java.net.URL;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
 import java.util.Scanner;
 
-import drjoliv.jfunc.collection.Unit;
+import drjoliv.jfunc.contorl.Either;
 import drjoliv.jfunc.contorl.Try;
+import drjoliv.jfunc.data.Unit;
 import drjoliv.jfunc.function.F1;
 
 public class IO {
@@ -31,6 +30,26 @@ public class IO {
     in.close();
     return buffer.toString();
   });
+
+  public static F1<InputStream,Try<String>> readFileFromStream = s -> Try.with(() -> {
+    System.out.println(s == null);
+    Scanner in = new Scanner(s);
+    StringBuffer buffer = new StringBuffer();
+    while(in.hasNextLine())
+      buffer.append(in.nextLine());
+    in.close();
+    return buffer.toString();
+  });
+
+  public static void main(String[] args) {
+    URL url = IO.class.getResource("hello.txt");
+    System.out.println(url == null);
+    Try<InputStream> stream = Try.with(() -> url.openStream());
+    Try<String> test = stream.bind(readFileFromStream);
+    Either<Exception,String> e = test.run();
+    e.consume(ex -> ex.printStackTrace()
+        , s -> System.out.println(s));
+  }
 
   public static Try<String> now = Try.with(() -> {
     return System.currentTimeMillis() + "";